LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-snap/scroll-snap-stop-002-nested-expected.txt


FAIL `scroll-snap-stop: always` snap point is further than the scroll destination and a snap area covers the snapport assert_equals: expected 900 but got 985
